refactor(nodes): extract core types and constants to dedicated crate
Strictly mechanical move of NodeExecResult struct and environment/SSH
constants from crates/gateway/src/node_exec.rs to new crate
crates/node-exec-types/. No logic changed.

Changes:
- New crate: moltis-node-exec-types (serde dependency only)
- Moved: NodeExecResult, SAFE_ENV_ALLOWLIST, SAFE_ENV_PREFIX_ALLOWLIST,
  BLOCKED_ENV_PREFIXES, SSH_ID_PREFIX, SSH_TARGET_ID_PREFIX
- Gateway now re-exports via pub use moltis_node_exec_types::{...}

Verification:
- All 8 node_exec tests pass
- just format-check passes
- 1:1 code migration verified via git diff
