+++ crates/gateway/Cargo.toml
@@ -36,6 +36,7 @@ moltis-memory          = { workspace = true }
 moltis-metrics         = { optional = true, workspace = true }
 moltis-msteams         = { workspace = true }
 moltis-network-filter  = { features = ["proxy", "service"], optional = true, workspace = true }
+moltis-node-exec-types = { workspace = true }
 moltis-nostr           = { optional = true, workspace = true }
 moltis-oauth           = { workspace = true }
 moltis-onboarding      = { workspace = true }
