+++ crates/node-exec-types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name                 = "moltis-node-exec-types"
+version.workspace    = true
+edition.workspace    = true
+repository.workspace = true
+
+[dependencies]
+moltis-metrics = { optional = true, workspace = true }
+serde          = { workspace = true, features = ["derive"] }
+tracing        = { workspace = true }
+
+[features]
+default = []
+metrics = ["dep:moltis-metrics"]
+
+[lints]
+workspace = true
